requirements.txt
pip~=20.3.3
websocket-client~=0.58.0
playsound~=1.2.2